api/requirements.txt
fastapi
uvicorn
yt-dlp
boto3
python-dotenv
pydantic